first commit system

+++ database/migrations/2026_07_10_103652_add_kuota_juara_to_lomba_table.php
@@ -6,24 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('lomba', function (Blueprint $table) {
-            // Menambahkan kolom kuota_juara, default 0 (Untuk penanda All Trophy)
-            $table->integer('kuota_juara')->default(0)->nullable();
+            if (!Schema::hasColumn('lomba', 'kuota_juara')) {
+                $table->integer('kuota_juara')->default(0)->nullable();
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('lomba', function (Blueprint $table) {
-            $table->dropColumn('kuota_juara');
+            if (Schema::hasColumn('lomba', 'kuota_juara')) {
+                $table->dropColumn('kuota_juara');
+            }
         });
     }
 };
